+++ src/core/audio/midi.hpp
@@ -6,8 +6,6 @@ struct MidiEvent {
 
   typedef unsigned char byte;
 
-  virtual ~MidiEvent() {}
-
   enum EventType {
     NOTE_OFF = 0b1000,
     NOTE_ON = 0b1001,
@@ -15,9 +13,7 @@ struct MidiEvent {
   } type;
 
   int channel;
-
   byte *data;
-
   int time;
 
 };
@@ -27,21 +23,21 @@ struct NoteOnEvent : public MidiEvent {
   int key = data[0];
   int velocity = data[1];
 
-  NoteOnEvent(MidiEvent event) : MidiEvent(event) {};
+  NoteOnEvent(const MidiEvent& event) : MidiEvent(event) {};
 };
 
 struct NoteOffEvent : public MidiEvent {
   int key = data[0];
   int velocity = data[1];
 
-  NoteOffEvent(MidiEvent event) : MidiEvent(event) {};
+  NoteOffEvent(const MidiEvent& event) : MidiEvent(event) {};
 };
 
 struct ControlChangeEvent : public MidiEvent {
   int controler = data[0];
   int value = data[1];
 
-  ControlChangeEvent(MidiEvent event) : MidiEvent(event) {};
+  ControlChangeEvent(const MidiEvent& event) : MidiEvent(event) {};
 };
 
 using MidiEventPtr = top1::SmartPolyPtr<MidiEvent,
