+++ src/core/faust-module.hpp
