+++ src/core/ui/mainui.hpp
@@ -1,10 +1,11 @@
 #pragma once
 
-#include "base.h"
+#include "core/ui/base.hpp"
 
 class MainUI : public ui::Screen {
 
   static void mainRoutine();
+  ui::
 
   bool globKeyPre(ui::Key key);
   bool globKeyPost(ui::Key key);
