+++ src/core/ui/waveform-widget.hpp
