+++ src/modules/synth-sampler.hpp
