+++ src/modules/tapedeck.cpp
@@ -1,15 +1,15 @@
-#include "tape.h"
+#include <bits/std_abs.h>
+#include <plog/Log.h>
+#include <sys/types.h>
+#include <cmath>
+
 #include "../audio/jack.h"
 #include "../globals.h"
-#include <plog/Log.h>
-#include <thread>
-#include <cmath>
-#include <algorithm>
-#include <mutex>
-#include <condition_variable>
-
 #include "../ui/mainui.h"
 #include "../ui/utils.h"
+#include "modules/../util/tapebuffer.h"
+#include "modules/metronome.h"
+#include "tape.h"
 
 using namespace top1;
